requirements.txt
requests
pandas
kafka-python
pyspark
matplotlib
jupyterlab
